Remove asm_bridge dependencies
