--- crates/vm/src/arch/asm_bridge/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "asm_bridge"
-version.workspace = true 
-edition.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[package.metadata.cargo-shear]
-ignored = ["cc"]
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-cc = "1.0"
-openvm-circuit = { workspace = true, features = ["test-utils"] }
-openvm-stark-sdk.workspace = true
-openvm-instructions.workspace = true
